1.删除冗余的sdl依赖库， 修改为环境变量的方式 2.修改支持64位操作系统 3.修改warning警告

+++ src/MediaCore/OutputMediaSource/direct3d_render.cpp
@@ -153,7 +153,7 @@ namespace output {
         rect.right = item.x_pos + item.width;
         rect.bottom = item.y_pos + item.height;
         const auto content = utils::CharacterSetConvertor::Instance()->UTF8ToUnicode(item.content);
-        font_->DrawText(nullptr, &content[0], content.length(), &rect,
+        font_->DrawText(nullptr, &content[0], static_cast<INT>(content.length()), &rect,
                         DT_SINGLELINE | DT_NOCLIP | DT_CENTER | DT_VCENTER, D3DXCOLOR(1.0f, 1.0f, 1.0f, 0.8f));
       }
     }
